Refactor kernel imports and remove CredentialProxy (#33)
- Updated import paths for ProtoKernel from `xulcan.kernel.runtime` to `xulcan.kernel.orchestrator` across multiple files.
- Removed CredentialProxy class from the public API, as runtime credential resolution is now handled during manifest materialization in `xulcan.runtime.resolver.ManifestResolver`.
- Adjusted related documentation and imports in the registry and system modules to reflect the removal of CredentialProxy.
