+++ app/xulcan/registry/__init__.py
@@ -4,20 +4,18 @@ This package provides infrastructure for dependency injection and credential man
 
 Key exports:
     - ProviderRegistry: Generic abstract factory for all adapters
-    - CredentialProxy: Injects credentials into adapters at build time
     - ToolSecretsVault: Vault for tool-specific credentials
     - RegistryContainer: Autocontained holder for all Xulcan registries (Issue 23)
     - bootstrap_registries: Pure function to populate RegistryContainer (Issue 23)
 """
 
 from .base import ProviderRegistry
-from .credentials import CredentialProxy, ToolSecretsVault
+from .credentials import ToolSecretsVault
 from .container import RegistryContainer
 from .bootstrap import bootstrap_registries
 
 __all__ = [
     "ProviderRegistry",
-    "CredentialProxy",
     "ToolSecretsVault",
     "RegistryContainer",
     "bootstrap_registries",
