+++ app/xulcan/system/registry.py
@@ -1,2 +1,2 @@
 # Deprecated: import from xulcan.registry instead
-from xulcan.registry import ProviderRegistry, CredentialProxy, ToolSecretsVault
+from xulcan.registry import ProviderRegistry, ToolSecretsVault
